Remove integration tests from coverage data
