+++ src/__tests__/check-sync_test.js
@@ -1,11 +1,9 @@
 // @flow
-import path from "path";
-import chalk from "chalk";
 import * as GetFiles from "../get-files.js";
 import * as GetMarkersFromFiles from "../get-markers-from-files.js";
 import * as HandleViolations from "../handle-violations.js";
+import * as CwdRelativePath from "../cwd-relative-path.js";
 import Logger from "../logger.js";
-import StringLogger from "../string-logger.js";
 
 import checkSync from "../check-sync.js";
 import ErrorCodes from "../error-codes.js";
@@ -13,16 +11,9 @@ import ErrorCodes from "../error-codes.js";
 jest.mock("../get-files.js");
 jest.mock("../get-markers-from-files.js");
 jest.mock("../handle-violations.js");
+jest.mock("../cwd-relative-path.js");
 
 describe("#checkSync", () => {
-    beforeAll(() => {
-        chalk.enabled = false;
-    });
-
-    afterAll(() => {
-        chalk.enabled = true;
-    });
-
     it("should expand the globs to files", async () => {
         // Arrange
         const NullLogger = new Logger();
@@ -212,6 +203,7 @@ describe("#checkSync", () => {
             "violation1",
             "violation2",
         ]);
+        jest.spyOn(CwdRelativePath, "default").mockImplementation(f => f);
         const errorSpy = jest.spyOn(NullLogger, "error");
 
         // Act
@@ -222,63 +214,5 @@ describe("#checkSync", () => {
             "checksync --fix violation1 violation2",
             true,
         );
-    });
-
-    it("should report __examples__ violations", async () => {
-        // Arrange
-        jest.spyOn(GetFiles, "default").mockImplementation((...args) =>
-            jest.requireActual("../get-files.js").default(...args),
-        );
-        jest.spyOn(HandleViolations, "default").mockImplementation((...args) =>
-            jest.requireActual("../handle-violations.js").default(...args),
-        );
-        jest.spyOn(GetMarkersFromFiles, "default").mockImplementation(
-            (...args) =>
-                jest
-                    .requireActual("../get-markers-from-files.js")
-                    .default(...args),
-        );
-        const stringLogger = new StringLogger();
-
-        // Act
-        await checkSync(
-            [path.join(__dirname, "../../__examples__")],
-            false,
-            ["//", "#"],
-            stringLogger,
-        );
-        const result = stringLogger.getLog();
-
-        // Assert
-        expect(result).toMatchSnapshot("__examples__");
-    });
-
-    it("should report __examples__ parse errors with autoFix", async () => {
-        // Arrange
-        jest.spyOn(GetFiles, "default").mockImplementation((...args) =>
-            jest.requireActual("../get-files.js").default(...args),
-        );
-        jest.spyOn(HandleViolations, "default").mockImplementation((...args) =>
-            jest.requireActual("../handle-violations.js").default(...args),
-        );
-        jest.spyOn(GetMarkersFromFiles, "default").mockImplementation(
-            (...args) =>
-                jest
-                    .requireActual("../get-markers-from-files.js")
-                    .default(...args),
-        );
-        const stringLogger = new StringLogger();
-
-        // Act
-        await checkSync(
-            [path.join(__dirname, "../../__examples__")],
-            true,
-            ["//", "#"],
-            stringLogger,
-        );
-        const result = stringLogger.getLog();
-
-        // Assert
-        expect(result).toMatchSnapshot("__examples__ autofix");
     });
 });
