+++ src/cwd-relative-path.js
@@ -1,0 +1,6 @@
+// @flow
+import path from "path";
+
+export default function(filePath: string) {
+    return path.relative(process.cwd(), filePath);
+}
