+++ src/report-violation.js
@@ -1,5 +1,6 @@
 // @flow
 import Format from "./format.js";
+import cwdRelativePath from "./cwd-relative-path.js";
 
 import type {ViolationHandler, ILog} from "./types.js";
 
@@ -24,7 +25,9 @@ const violationReporter: ViolationHandler = function(
     fixable: boolean,
     log: ILog,
 ): void {
-    log.log(Format.violation(`${sourceFile}:${sourceLine + 1}:0`));
+    log.log(
+        Format.violation(`${cwdRelativePath(sourceFile)}:${sourceLine + 1}:0`),
+    );
     // TODO: Can we sync a marker within the same file? Not really since there could be multiple;
     //       how would we manage that? Let's ban it.
 };
